requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cross-web==0.4.1
cryptography==46.0.5
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.128.6
graphql-core==3.2.7
greenlet==3.3.2
h11==0.16.0
idna==3.11
packaging==26.0
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.13.0
pyjwt==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==7.1.1
requests==2.32.5
six==1.17.0
sqlalchemy==2.0.48
starlette==0.52.1
strawberry-graphql==0.294.0
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uvicorn==0.40.0
